+++ src/main/java/com/voucher/vouchermanagement/manager/io/VoucherManagerConsole.java
@@ -1,4 +1,4 @@
-package com.voucher.vouchermanagement.utils.io.manager;
+package com.voucher.vouchermanagement.manager.io;
 
 import com.voucher.vouchermanagement.manager.command.CommandType;
 import com.voucher.vouchermanagement.model.voucher.VoucherType;
