+++ src/main/java/com/voucher/vouchermanagement/manager/io/VoucherManagerInput.java
@@ -1,6 +1,5 @@
-package com.voucher.vouchermanagement.utils.io.manager;
+package com.voucher.vouchermanagement.manager.io;
 
 public interface VoucherManagerInput {
-
     String input(String prompt);
 }
